move init_encrypted_tunnels to init_tunnel.rs
