+++ src/init_tunnel.rs
@@ -2,11 +2,12 @@ use std::fmt::Debug;
 
 use serde::{Deserialize, Serialize};
 use serde_json::json;
-use wasm_bindgen::{JsValue, UnwrapThrowExt, prelude::wasm_bindgen};
+use wasm_bindgen::{JsValue, prelude::wasm_bindgen, UnwrapThrowExt};
 use web_sys::console;
 
 use ntor::client::NTorClient;
 use ntor::common::{InitSessionResponse, NTorCertificate, NTorParty};
+use std::rc::Rc;
 
 use crate::constants::MAX_INIT_TUNNEL_ATTEMPTS;
 use crate::utils;
@@ -14,6 +15,9 @@ use crate::{
     http_call_indirection::{HttpCaller, HttpCallerResponse},
     network_state::DEV_FLAG,
 };
+use crate::http_call_indirection::ActualHttpCaller;
+use crate::network_state::{NETWORK_STATE, NetworkState, NetworkStateOpen, ServiceProvider};
+use crate::utils::base_url;
 
 #[derive(Clone)]
 #[wasm_bindgen(getter_with_clone)]
@@ -60,7 +64,8 @@ impl Debug for InitTunnelResult {
 pub async fn init_tunnel(
     backend_url: String,
     http_caller: impl HttpCaller,
-) -> Result<InitTunnelResult, JsValue> {
+) -> Result<InitTunnelResult, JsValue>
+{
     let dev_flag = DEV_FLAG.with_borrow(|flag| *flag);
 
     // 1. Initialize NTor Client message
@@ -165,4 +170,65 @@ pub async fn init_tunnel(
     };
 
     Ok(result)
+}
+
+/// This function initializes the encrypted tunnel for the given service providers using a background process, which updates
+/// the `NETWORK_STATE` global static.
+#[wasm_bindgen(js_name = "initEncryptedTunnel")]
+pub fn init_encrypted_tunnels(
+    forward_proxy_url: String,
+    service_providers: Vec<ServiceProvider>,
+    dev_flag: Option<bool>,
+) -> Result<(), JsValue> {
+    if let Some(val) = dev_flag {
+        if val {
+            DEV_FLAG.with_borrow_mut(|flag| *flag = true);
+            console::log_1(&"Dev mode enabled".into());
+        }
+    }
+
+    for service_provider in service_providers {
+        // update the urls as connecting before scheduling the background task to initialize the tunnel
+        NETWORK_STATE.with_borrow_mut(|cache| {
+            cache.insert(
+                service_provider.url.clone(),
+                Rc::new(NetworkState::CONNECTING),
+            );
+        });
+
+        let base_url = base_url(&service_provider.url)?;
+        let backend_url = format!("{}/init-tunnel?backend_url={}", forward_proxy_url, base_url);
+        let forward_proxy_url = forward_proxy_url.clone();
+
+        // schedule the background task to initialize the tunnel
+        wasm_bindgen_futures::spawn_local(async move {
+            match init_tunnel(backend_url, ActualHttpCaller).await {
+                Ok(val) => {
+                    if dev_flag.unwrap_or(false) {
+                        console::log_1(
+                            &format!("Tunnel initialized for {}", service_provider.url).into(),
+                        );
+                    }
+
+                    let state = NetworkStateOpen {
+                        http_client: reqwest::Client::new(),
+                        init_tunnel_result: val,
+                        forward_proxy_url: forward_proxy_url.clone(),
+                    };
+
+                    NETWORK_STATE.with_borrow_mut(|cache| {
+                        cache.insert(base_url, Rc::new(NetworkState::OPEN(state)));
+                    });
+                }
+
+                Err(err) => {
+                    NETWORK_STATE.with_borrow_mut(|cache| {
+                        cache.insert(base_url, Rc::new(NetworkState::ERRORED(err)));
+                    });
+                }
+            }
+        });
+    }
+
+    Ok(())
 }
